# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ever Works Website Template API",
    "description": "REST API documentation for the Ever Works website template. This template provides a complete foundation for building modern web applications with authentication, user management, favorites, and content management features.",
    "version": "1.0.0",
    "contact": {
      "name": "Ever Works Template",
      "url": "https://demo.ever.works"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://demo.ever.works",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/admin/roles/stats": {
      "get": {
        "tags": ["Admin - Roles"],
        "summary": "Get role statistics",
        "description": "Returns comprehensive statistics about roles including total count, breakdown by active/inactive status, and average permissions per role. Data is sourced from database. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Role statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoleStatsResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "total": 8,
                    "active": 6,
                    "inactive": 2,
                    "averagePermissions": 12.5
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Forbidden"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/admin/tags": {
      "get": {
        "tags": ["Admin - Tags"],
        "summary": "List tags (Admin)",
        "description": "Returns a paginated list of tags. Tags are loaded from Git repository. Supports pagination with page and limit parameters. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of tags per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Tags retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTagsResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "tags": [
                      {
                        "id": "productivity",
                        "name": "Productivity",
                        "isActive": true,
                        "count": 45
                      },
                      {
                        "id": "design",
                        "name": "Design",
                        "isActive": true,
                        "count": 32
                      },
                      {
                        "id": "development",
                        "name": "Development",
                        "isActive": false,
                        "count": 28
                      }
                    ],
                    "total": 156,
                    "page": 1,
                    "limit": 10,
                    "totalPages": 16
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Admin - Tags"],
        "summary": "Create tag (Admin)",
        "description": "Creates a new tag in the Git repository. Validates tag ID format (lowercase, hyphens, numbers only) and name length. Checks for duplicates. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagRequest"
              },
              "example": {
                "id": "machine-learning",
                "name": "Machine Learning",
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTagResponse"
                },
                "example": {
                  "success": true,
                  "tag": {
                    "id": "machine-learning",
                    "name": "Machine Learning",
                    "isActive": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missing_fields": {
                    "summary": "Missing required fields",
                    "value": {
                      "success": false,
                      "error": "Tag ID and name are required"
                    }
                  },
                  "invalid_id": {
                    "summary": "Invalid ID format",
                    "value": {
                      "success": false,
                      "error": "Tag ID must contain only lowercase letters, numbers, and hyphens"
                    }
                  },
                  "invalid_length": {
                    "summary": "Invalid length",
                    "value": {
                      "success": false,
                      "error": "Tag name must be between 2 and 50 characters"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "Conflict - tag already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Tag with this ID already exists"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/admin/tags/{id}": {
      "get": {
        "tags": ["Admin - Tags"],
        "summary": "Get tag by ID (Admin)",
        "description": "Returns a specific tag by its ID. Data is loaded from Git repository. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the tag",
            "example": "productivity"
          }
        ],
        "responses": {
          "200": {
            "description": "Tag retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTagResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "productivity",
                    "name": "Productivity",
                    "isActive": true
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Tag not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Tag not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "tags": ["Admin - Tags"],
        "summary": "Update tag (Admin)",
        "description": "Updates a tag's name and/or active status. Tag ID cannot be changed. Data is stored in Git repository. Validates name length and checks for duplicates. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the tag",
            "example": "productivity"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagRequest"
              },
              "example": {
                "name": "Enhanced Productivity",
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTagResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "productivity",
                    "name": "Enhanced Productivity",
                    "isActive": true
                  },
                  "message": "Tag updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missing_name": {
                    "summary": "Missing name",
                    "value": {
                      "success": false,
                      "error": "Tag name is required"
                    }
                  },
                  "invalid_length": {
                    "summary": "Invalid length",
                    "value": {
                      "success": false,
                      "error": "Tag name must be between 2 and 50 characters"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Tag not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Tag not found"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - tag name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Tag with this name already exists"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": ["Admin - Tags"],
        "summary": "Delete tag (Admin)",
        "description": "Deletes a tag from the Git repository. This is a permanent deletion. Requires admin privileges.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique identifier of the tag",
            "example": "productivity"
          }
        ],
        "responses": {
          "200": {
            "description": "Tag deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTagResponse"
                },
                "example": {
                  "success": true,
                  "message": "Tag deleted successfully"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Tag not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Tag not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Session-based authentication using NextAuth.js"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Additional error details"
          }
        },
        "required": ["success", "error"]
      },
...
      "AdminRoleStatsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/RoleStats"
          }
        },
        "required": ["success", "data"]
      },
      "RoleStats": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total number of roles in the system",
            "example": 8
          },
          "active": {
            "type": "integer",
            "description": "Number of roles with active status",
            "example": 6
          },
          "inactive": {
            "type": "integer",
            "description": "Number of roles with inactive status",
            "example": 2
          },
          "averagePermissions": {
            "type": "number",
            "description": "Average number of permissions per role (rounded to 1 decimal place)",
            "example": 12.5
          }
        },
        "required": ["total", "active", "inactive", "averagePermissions"]
      },
...
      "AdminTagsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/TagListResponse"
          }
        },
        "required": ["success", "data"]
      },
      "TagListResponse": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagWithCount"
            },
            "description": "Array of tags with optional count information"
          },
          "total": {
            "type": "integer",
            "description": "Total number of tags",
            "example": 156
          },
          "page": {
            "type": "integer",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Number of tags per page",
            "example": 10
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages",
            "example": 16
          }
        },
        "required": ["tags", "total", "page", "limit", "totalPages"]
      },
      "TagWithCount": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique tag identifier (lowercase, hyphens, numbers only)",
            "example": "productivity"
          },
          "name": {
            "type": "string",
            "description": "Display name of the tag",
            "example": "Productivity"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the tag is active",
            "example": true
          },
          "count": {
            "type": "integer",
            "nullable": true,
            "description": "Number of items using this tag (optional)",
            "example": 45
          }
        },
        "required": ["id", "name", "isActive"]
      },
      "CreateTagRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "pattern": "^[a-z0-9-]+$",
            "description": "Unique tag identifier (lowercase, hyphens, numbers only)",
            "example": "machine-learning"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Display name of the tag",
            "example": "Machine Learning"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the tag is active",
            "default": true,
            "example": true
          }
        },
        "required": ["id", "name"]
      },
      "CreateTagResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "tag": {
            "$ref": "#/components/schemas/TagData"
          }
        },
        "required": ["success", "tag"]
      },
      "TagData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique tag identifier",
            "example": "machine-learning"
          },
          "name": {
            "type": "string",
            "description": "Display name of the tag",
            "example": "Machine Learning"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the tag is active",
            "example": true
          }
        },
        "required": ["id", "name", "isActive"]
      },
      "GetTagResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/TagData"
          }
        },
        "required": ["success", "data"]
      },
      "UpdateTagRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "description": "Updated display name of the tag",
            "example": "Enhanced Productivity"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the tag is active",
            "example": true
          }
        },
        "required": ["name"]
      },
      "UpdateTagResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/TagData"
          },
          "message": {
            "type": "string",
            "example": "Tag updated successfully"
          }
        },
        "required": ["success", "data", "message"]
      },
      "DeleteTagResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Tag deleted successfully"
          }
        },
        "required": ["success", "message"]
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Unauthorized"
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "Internal server error"
            }
          }
        }
      }
    }
  }
}